requirements.txt
streamlit
numpy
pandas
Pillow
tensorflow-cpu